Allow workspace editors except guest role in team to add/remove members from workspace
