+++ .vscode/settings.json
@@ -1,11 +1,13 @@
 {
   "cSpell.words": [
     "alphanum",
+    "clsx",
     "deduping",
     "DICEBEAR",
     "gettask",
     "humburger",
     "Kanban",
+    "listview",
     "lucide",
     "NEXTAUTH",
     "orgnise",
