+++ src/lib/swr/use-workspaces.ts
@@ -21,6 +21,7 @@ interface IWorkspaces {
   mutate: KeyedMutator<any>;
   loading: boolean;
   workspaces: Workspace[];
+  activeWorkspace?: Workspace;
   deleteWorkspace: (workspaceSlug: string) => Promise<void>;
   updateWorkspace: ({ name, description, slug, visibility, defaultAccess }: UpdateWorkspace, teamSlug: string) => Promise<void>;
 }
@@ -78,8 +79,14 @@ export default function useWorkspaces(): IWorkspaces {
   // Delete workspace
   async function deleteWorkspace(workspaceSlug: string) { }
 
+
+  const activeWorkspace = data?.workspaces?.find(
+    (w: any) => w?.meta?.slug === workspace_slug,
+  );
+
   return {
     mutate,
+    activeWorkspace,
     workspaces: data?.workspaces,
     error,
     loading: !data && !error ? true : false,
